(PE-21788) Open port 8170 on ec2 masters
The pe_acceptance_tests suite hits the code manager endpoint on primary
masters at 8170 during test runs. Prior to this patch, the ec2_helper
was generating a security group for master nodes which did not include
port 8170, and the
https://github.com/puppetlabs/pe_acceptance_tests/blob/2017.3.x/acceptance/tests/code-manager/deploy_with_code_manager_API.rb
test was failing when Scooter could not reach master:8170 from the beaker test
runner.

The patch just adds 8170 to the list of open ports on a master node.