BKR-1487: Custom CIDRs for security groups
Without this changes, we encounter several issues:
- In the case of setting a custom 'vpc_id' and 'subnet_id' in the host
the ping security group gets created in the correct network and the other one in the
default vpc which is not desirable as the EC2 won't be able to use it

- Also given the security requirements in some companies using 0.0.0.0/0
might introduce security risks so being able to set CIDRs mitigates such
issues.

This commit introduces 'sg_cidr_ips' which can be added as a host
property and is a comma seperated list of CIDRs (e.g. 192.168.0.1/24,10.0.0.10/24)
I guess this might allow for adding other security groups as well.

Finally, replace 'security_groups' with 'security_group_ids' as AWS create
instance API call complains about groupName when the none default VPC is used.
A manual test has been made to make sure 'security_group_ids' works with default
VPC as well.