BKR-1522 Add options to drop some of the provisioning
'disable_root_ssh' in case we don't want to configure root user

This is the case when using beaker with a Windows AMI which has openssh server
with a public key already set in the Administrator user. In that case beaker only needs
to import the public key to AWS and use it for ssh.

Also some of the provision steps namely set_hostnames and configure_hosts are not
compatible with windows and it's shell (cmd or Powershell). So this stops those
steps from running on windows platform.