(maint) Fix BKR-1546 for the case where subnet is nil
Before this change the tests would run launch_all_nodes() and call create_instance()
with the subnet set to 'nil' by design. The sdk accepts a node with subnet_id = 'nil'
but not network_interfaces.subnet_id = 'nil' so the replacement done in BKR-1546
where we always create a 'network_interfaces' (with or without a public ip) is not
functionally the same.