(PDK-1275) tagging and assigning correct security groups
Previously when it tried to tag after waiting for netdev status it was too late as the instances already started terminating through the AWS policy. By tagging them sooner it gives all instances a chance to report as `OK`.

Same with the assigning of security groups, as during the create_instances it was too early to have a `subnet_id` or a `assoc_pub_ip_addr` which meant the security groups which were created never got assigned. So similar to tagging, the network_interface will be modified at an appropriate stage.